Optimized PDF reference parsing
Initially the reference parsing code was moved from #next_token
to #parse_object since only in the latter context are references
useful. This provided a small performance increase.

However, the resetting of the parse position due to the checking
for "int int R" leads to very bad performance. The solution is
to use a simplified (meaning not 100% spec compatible) reference
checking mechanism.

The new solution provides a performance increase between 20% and
100% depending on the parsed file.

The comprise is that there *may* be PDF files that use a comment
somewhere in between the parts of a PDF reference. However, the
chance for this is rather small and I haven't encountered one so
far.