Optimization: Avoid unnecessary string allocations in Stream
Document#wrap creates a new HexaPDF::PDF::Object. However,
when a Stream was created, the stream argument of
Stream#initialize was not set and therefore an empty string
was first assigned as stream data. Afterwards the correct
stream data was assigned using Stream#stream=.

By reordering the code this unnecessary string allocation is
now removed.