Refactored internals of PDF objects
Introduced a new internal value object PDFData which holds
the value, oid, gen and stream data of one PDF object. A
PDF object is now really just a wrapper around such a PDFData
object.

This fixes the optimization task to work correctly because
before objects were stored in a revision and may have been
re-wrapped when found inside another object. This rewrapping
led to the oid/gen being changed in only one of the objects
and not the other.