Updated Dictionary validation to validate nested direct PDF objects
The former check just validated values of the Dictionary itself.
However, this may not be enough when the tree was built manually.
Therefore if the value of a key is a hash, it is recursively
checked for direct PDF objects which are then validated.

The check for PDFByteString having a binary encoding is not
necessary anymore and has been removed (this is already done by
the PDFByteStringConverter).