Refactored color space implementation
The previous way of representing a color space through a module only
works for the device color spaces because there is only one DeviceGray,
one DeviceRGB and one DeviceCMYK color space. However, there are, for
example, many different CalRGB color spaces.

Therefore the color spaces are now all classes, with a special
implementation for the device color spaces.

Additionally, the Processor object doesn't handle the color spaces any
more. This will be done by the resource dictionary in the future since
custom color spaces are defined there. To facilitate this the mapping of
color space name to implementation is now managed by the global
configuration object.