Removed class level object validators in favor of inheritance + super
The new validation scheme uses the inheritance chain and a call to super
to accomplish validation instead of HexaPDF::Object.define_validator.

One small disadvantage of the new scheme might be that the super call
has to be done. However, an advantage is that custom validation can now
be added to a class or even a single object by defining a module and
including it in the class or extending that object.