Optimized writing by caching the object to object stream mapping
Previously, when writing an object of a revision, each object stream was
asked if it contained that object. In case of many objects and many
object streams this leads to bad performance (O(n * m)).

The new implementation uses a lookup table that tells which object was
written to which object stream (O(n)).