Added message listener interface to Document
There was no way to run some code when a certain event occured. However,
this is needed in some cases. For example, some PDF objects may need to
finalize some work before being written.

This could potentially be done in the validator callback but that is not
its purpose.

Also, two messages are now dispatched:
* :complete_objects when a document should be written but before any
  step like validation is actually taken. This message can also be
  called any time and should not have any unwanted side effects, i.e.
  calling it multiple times should always result in the same state
  afterwards.
* :before_write directly before writing of a PDF document occurs