Added generic font loading support
Font loading is done similar to image loading. Using the configuration
font_loader it is possible to add code that handles loading of fonts,
e.g. there could be a font loader for Linux that uses fontconfig for
getting the font files.

There is now a Document#fonts method that returns a utility object that
handles font related stuff like font loading.