Updated Canvas implementation to close open graphics objects on serialization
The Canvas class provides easy access to all PDF drawing operations and
makes sure that each operation can only be used in the correct way.
However, until this commit it was possible to leave the content stream
in an invalid state at serialization since there might be an open
path/text object or not restored graphics states.

By using a StreamData object and not a plain String it is now ensured
that the content stream is always correct.