Implemented a CMap writer for writing ToUnicode CMaps
This was not necessary for the standard 14 PDF fonts since there we
could rely on the /Encoding and the Adobe GlyphList for a code to
Unicode codepoint mapping.

However, in the context of composite fonts it is necessary because it
allows using the Identity-H encoding for the character code to CID
mapping and the Identity CID to GID mapping which makes, for example,
using TrueType fonts much easier.