Refactored font support in Canvas for better performance
The Canvas#show_glyphs needed to check the font dictionary on every
encoded glyph. Since the dictionary is always the same for TrueType
fonts, this is not really needed in most cases.

Therefore each wrapped font now supports only one font dictionary. This
is only a restriction for the built-in standard 14 PDF fonts. However,
we assume that in most cases TrueType fonts are used and therefore this
small inconvenience is okay.