Change optimization argument and default value of Document#write
Since the method defaulted to generating object streams, one had to
take special care if that was not wanted. For example, when
optimization had already be done or if a PDF with a version lower
than 1.5 should be created.

So rename the argument to 'optimize' and set its default value to false.