Fix bug in security handler
The standard security handler uses the attribute EncryptMetadata which
utilizes a default value. If this key is not set, the default value is
automatically assigned when accessed during decryption, invalidating the
hash used for checking whether the encryption dictionary is still valid.

The solution is to compute the hash after the complete decryption setup
has been done.