Handle potentially indirect objects correctly in HexaPDF::Object.deep_copy
A PDF object that is not yet indirect but will be after validation, must
not be deep copied.

This fixes amongst other things a problem with infinite recursion, e.g.
when an as yet non-indirect /Page object refers to its non-indirect
parent /Pages object which refers back to the /Page object.