Update 'hexapdf modify' command to support files with many pages
Testing a file with 28000 pages lead to performance problems because
each access of a page with a certain index needed to go through all
prior pages, giving a complexity of O(n^2/2).

Create an array of all pages once and use that for accessing the pages,
reducing the complexity to O(2n).