Fix writing of hybrid-reference PDF files
The easy fix is to just don't write the /XRefStm key-value pair when
outputting. However, this potentially leads to bigger output files when
the cross-reference table and the cross-reference stream referenced by
/XRefStm refer to the same, identical objects (i.e. objects with the
same position in the file) because they are written twice.

To avoid this the entries from the cross-reference stream are used to
update the entries in the corresponding cross-reference table. I.e. the
table referred to by the /Prev entry is merged with the one from the
/XRefStm entry to provide a single revision, not two.

With this change we loose the functionality to support /XRefStm entries
without accompanying /Prev entries. However, this is something that
conforming PDF writers should not produce.

This also means that hybrid-reference files are written as standard PDF
1.5+ files.