Add methods for caching expensive computations of PDF objects
Some PDF objects need to cache expensive operations. For example, font
types need to cache the parsed ToUnicode CMap.

These cached values are currently held in instance variables. So if a
PDF object is wrapped with another type, the expensive operations must
be executed again.

Change the caching strategy to store the cached objects centrally in the
document instance.