Add Layout::InlineBox and Layout::LineFragment
A LineFragment describes a line of text made up of TextFragment objects
and provides methods for determining various metric values similar to
TextFragment.

In addition to TextFragment objects, the new InlineBox objects can be
also be used with a LineFragment to provide fixed size inline graphics.

The TextFragment class was adjusted to provide the additionally
necessary methods for use by LineFragments objects.