Implement an InvalidGlyph class for more flexible glyph handling
The current implementation raises an error when a Unicode string cannot
be fully converted to an array of glyphs. This is obstructive because
sometimes one wants to deal with invalid glyph mappings later on.

For example, there are generally no glyph mappings for CR and LF. This
would lead to an error early on, before advanced text layouting
facilities could detect the invalid glyphs and handle them properly,
e.g. by discarding them and breaking the line at the position.

Implement a new class that represents InvalidGlyph objects. This also
needs changes to the standard glyphs so that all Glyph/InvalidGlyph
objects can return their string representation, for error handling or
layouting decisions later on. And only raise an error if a user wants to
encode InvalidGlyphs, not when they are created.