Fix deletion of valid page nodes when optimizing page trees
When a PDF file contains a page with the same oid/gen in multiple
revisions, only the latest object should be retained when optimizing the
page tree with the "optimize" CLI command. This leads to an error
because all revisions of the page are deleted.

Fix this by only deleting the unused page nodes in their respective
revision.