Fix Stream#stream_source to always return a decrypted stream
The decryption/encryption should happen in a way sothat one doesn't need
to think about it. This is already true for most parts of HexaPDF.

The problem is that for encrypted streams Stream#stream_source returns
an undecrypted stream and Stream#stream_decoder returns a fully
decrypted *and* decoded stream. However, sometimes it is necessary to
retrieve just the undecrypted stream (e.g. when exporting a JPEG image
we don't want to decode the image data into pixels but just use the
encoded data itself).

Fix this by introducing a subclass of StreamData that is used for
encrypted streams.