Update filter predictor to handle common problems
Introduce a new global configuration option filter.predictor.strict that
specifies whether the predictor filter should raise an error or try to
handle problems in certain situations.

The addressed problem is that if the predictor misses some data for the
last row, assume that the row consists of a reduced number of bytes and
only process those.