Fix one-off error in LZWDecode
We are looking at the table size in each decode round and we know that
after the table is filled, i.e. the last entry 4095 ist assigned and the
table size gets 4096, a clear table code is emitted.

This means that on also decoding once the entry 4095 gets assigned, the
table has a size of 4096 and not 4095. So use 4096 for the check!