Optimize TrueTypeWrapper#encode by avoiding expensive hashing
The implementation of the #encode method already memoizes the result of
the operation but a lookup is still needed. This lookup was done using
the Glyph objects themselves. It turns out that getting the hash of such
a Glyph object is quite expensive, especially so because this method is
called very often.

Reduce the processing power needed for hashing by using the glyph ID as
hash key.