Update TextLayouter to support arbitrarily shaped areas
The current implementation of TextLayouter support varying line widths
and offsets from the left. However, it does not support lines with holes
in them.

To generalize the text layouting algorithm to support line with holes we
modify the width argument: Widths can be specified as a simple number or
an array of the form [offset, width, offset, width, ...]. The new array
form allows for an initial offset from the left and holes in the line.
Together with a block that varies the numbers in such arrays the layout
of text inside or around arbitrarily shaped arrays can be done.