Add Layout::Frame for easier box positioning and text layouting
This new class will be the basis for positioning arbitrary boxes and
also allows easy text layouting inside a polygon.

For working with polygons a new dependency on the 'geom2d' library is
added. This library defines the needed geometric objects and an
algorithm for polygon intersection. Such an algorithm is needed for
removing parts of a polygon, i.e. when a picture is placed inside a
frame.