Add Type::FontType3 for Type 3 font support
For this to work the Type::Font and Type::SimpleFont classes have to be
changed to remove things that are not supported by Type 3 fonts or
handled differently.

This should also fix some problems with validation errors since Type 3
fonts don't have a /BaseFont key but the /BaseFont key was defined in
the Type::Font ancestor class.