Fix TextLayouter::Result#height to respect first line offset
Don't use a provided height value but calculate the height from the
result lines themselves

The TextLayouter::Result#height method should return the needed height
for the whole result. As indicated by the API documentation this
includes a possible offset for the first line, for example, in case of
bottom or center vertical alignment.