Change Layout::Box#fit return value
The current implementation returns the width/height pair. This
information is also available through the Box#width/Box#height
accessors, so returning it is not necessary.

Return a boolean value instead of the width/height pair that signifies
whether the fitting was successful. This new method signature also
allows for subclasses that don't fit into a rectangular area.