Allow HexaPDF::Object and HexaPDF::Reference to be compared
Objects of both classes can already be used as interchangable hash keys.
This is useful because sometimes during processing we get reference and
sometimes objects.

By also allowing comparison between Object and Reference it will be
possible to compare freshly loaded indirect objects to possibly updated
ones (this happens, for example, when a Reference is resolved to an
Object during access).

The only thing we have to make sure is that oid/gen comparison is done
solely when comparing Object and Reference, and not Object and Object
since in the later case we have more information (the data object) and
can do a more thorough comparison.