Add Layout::Box#draw_content for subclassing purposes
All subclasses of Box will have custom drawing functionality and
currently rely on the draw_block block. This has the disadvantage that
an extra layer of the graphics stack is saved due to the need to
translate the origin to the bottom-left corner in the general case.

However, many subclasses can easily work from an arbitrary bottom-left
corner, with TextBox being the prime example.