Fix fitting of absolutely positioned boxes for Layout::Frame
The Box#fit method is not called on absolutely positioned boxes when
fitting is performed for a frame. However, the call sequence always has
to be fit, split (optional) and draw.

Since absolutely positioned boxes can have arbitrary size but still may
need some kind of fitting (e.g. for images when only a width is set
sothat the height is automatically calculated), we have to choose the
available space for them somehow. This is done by using the space to the
right and above the bottom left corner inside the frame.