Fix bug in variable width line wrapping due to line spacing
When checking whether a line still fits into the available height, we
calculate it by using the current height and adding the line's height
and the gap between the lines. However, the variable width line wrapping
calculations just use the current height plus the line's height for
determining the available width.

To fix this we need to yield not just the line height to the width block
supplied to the simple line wrapping algorithm, but a Line-like object
so that the line spacing gap can be correctly calculated and respected
in TextLayouter#fit.