Make the result of fitting a box in a frame available
The Frame class currently allows to fit a box and then split/draw it. To
do this the necessary state is kept by the Frame instance which means
that it is invalidated if another box object is provided.

To fit multiple boxes without rendering the state is extracted from the
Frame class and made available.