Freeze items when splitting a text fragment in Layout::TextLayouter#fit
If a too-wide text fragment is split, each glyph gets its own text
fragment. And when assembling a line, text fragments with the same style
get concatenated together, reusing unfrozen text fragment items if
possible.

Freeze the items of the split text fragments to make sure the items
don't get changed when used in a line.