Work around bug in Adobe Acrobat wrt object streams in encrypted files
The Catalog object may be in an object stream according to the PDF
specification. The only exception is for linearized documents where it
must be a normal indirect object.

However, Adobe Acrobat appears to have a bug because it doesn't accept
encrypted files in which the Catalog is in an object stream. Work around
this Acrobat bug by not compressing the Catalog object in such a case.