Fix loading and saving encrypted and signed PDFs
When a PDF gets signed, the signature is written after the complete PDF has
been created. This works by reserving space for the signature beforehand.
Due to this the signature which is stored as byte string is not encrypted but
HexaPDF tried to decrypt it anyways.

Handle the /Contents key of the /Sig dictionary specially to avoid this
problem.

Note: This is only a quick fix since implementing correct signature support
requires more things.