Fix Document#wrap to better handle subtype mappings without a type
The Document#wrap method allows mapping an object to a certain class using the
/Type and /Subtype (/S) keys. In many cases the PDF specification states that
the /Type key is optional.

To make this work nonetheless it was assumed that /Subtype (/S) keys are also
unique (as are the /Type keys). However, this is not the case and lead to
mappings of objects to invalid classes.

To prevent such invalid mappings in case the /Type key is not set but
/Subtype (or /S) is, we now also require that all required fields have values.
This should provided enough differentiation to always map to the correct
class.