Completely revamp the hexapdf inspect command
There are now two ways to do things with the inspect command, either
with the interactive mode or with the command line options.

Since the command line options are inferior to the interactive mode,
replace them with arguments on the command line that are interpreted as
interactive mode commands.

Furthermore, allow multiple commands to be specified on one line.