Fix line spacing of empty lines when using the text layouter
When using the Layout::TextLayouter class, each TextFragment provided to
the #fit method as well as the layouter itself has an associated style.

Until now the style of the text layouter was used for empty lines,
leading to problems when no font was set or when the font was different
from that of the text fragments.

To remedy the situation the penalty objects describing line breaks
inherit the style of the original text fragment.

In case the items provided to #fit are already segmented and penalty
objects describing line breaks do not have an item with a style
associated, the current behaviour is retained.