Add HexaPDF::PDFArray to support automatic reference resolution
HexaPDF had a special hash wrapping class HexaPDF::Dictionary from the
start to support various things like automatic reference solution and
field definitions. However, no special support for the other container
type array was added.

Due to the lack of this, special care had to be taken to make sure that
references were taken into account when processing arrays. In case of
the Rectangle class this was not done, leading to PDF processing
failures.

To avoid this, we introduce a new wrapper class HexaPDF::PDFArray for
arrays that works similar to Dictionary. Additionally, a new dictionary
field conversion module for PDFArray is added and all field definitions
are updated to use PDFArray instead of Array. The Rectangle class is now
also based on the new PDFArray.

The name PDFArray is used so that there are no conflicts with the
built-in Ruby Array class.