Update Utils::BitField to allow customizing bits in subclasses
There are situations where a bit field is defined in a super class and
subclasses need to extend the allowed bits.

The way to make this happen is by storing the allowed bits in a constant
in the super class and use the constant lookup of Ruby to choose the
right bits. I.e. if the constant is not defined in a subclass, the one
from the superclass is automatically used. But if the constant is
defined in the subclass, it is preferred.