Fix memory consumption problem when using PNG images with alpha channel
For PNGs with alpha channel we need to separate RGB information from the
alpha channel information. This needs to be done for each pixel of the
image.

The current code uses String#byteslice to do this in one go. However,
this also leads to two string allocations per pixels. Avoid the string
allocations by using String#getbyte and manual iteration.

Furthermore, don't explicitly delete each processed row from the image
data string but just use the rest. This also leads to fewer allocations
because due to an internal MRI optmizationl