Lazily parse the data of the hmtx tables of TrueType fonts
The current code parses all width and left side bearing information
up-front even though most of it is never used. This is especially
bad (performance and memory wise) for fonts with many glyphs like CJK
fonts.

Change the code to lazily parse the data for a specific glyph ID on
access.