Fix variable width layouting with a too wide text box
When one text box doesn't fit on a line, all the space below the current
height needs to be searched for possible space when using variable width
layouting.

If it happens that the text box would fit at the current location if the
line height is not considered, it is wrongly placed there because the
stored line height is not reset, leading to a wrong width calculation.

Fix this by reseting the stored line height.