Change color space and color implementations wrt to normalization
Only apply the normalization of values when using the ::color helper
method of the color space instead of always.

This allows creating a color object without normalization applied which
is useful, for example, when user provided input is given in already
normalized form (regardless of the class of the input values).