Remove normalization when invoking color operators
When a color value is set on the graphics state by invoking a color
operator, we have to assume that the value is already normalized. This
is because operators work directly on their values, so the values must
conform to the PDF spec.

Until now the #color method was used which does normalization. Switching
to the #prenormalized_color method makes sure that no normalization
happens.