Allow using a block with Content::Processor#parse
The #parse method currently needs a Processor object to work correctly.
However, sometimes we just want to parse a content stream quickly
without, for example, maintaining a correct graphics state.

To allow this, the #parse method now accepts a block that can be used
instead of the processor argument to handle the parsed PDF objects.