Fix Layout::TextBox to respect the set width/height on fitting/splitting
When the width or the height is set to anything other than zero on
HexaPDF::Layout::TextBox, it correctly lays out the text within these
bounds but the width/height calculation is wrong.

Furthermore, when width/height is set and bigger than the available
width/height, the box mustn't fit, no matter what.

This fixes both issues.