Fall back to the internal encoding of Type1 fonts when decoding strings
Type1 fonts with character set name of "Special" either don't use
standard glyph names (like the ZapfDingbats font) or map Unicode
codepoints to different names (like the Symbol font).

So when decoding a string like "ABC", the ZapfDingbats font will return
only missing glyphs because the glyph list for ZapfDingbats doesn't
contain the respective names.

However, the codepoints for A, B and C have a valid mapping in the
font's internal encoding. By using this mapping we get valid glyph
objects and can more easily use the ZapfDingbats font (otherwise we
would have to use the correct codepoints for these glyphs which are not
easily entered).

A similar story is for the Symbol font. In this case the glyph list has
the names for "ABC" but the font has no mapping to glyphs for these
names.

In such cases we fall back to the internal encoding. This allows use to
use the glyphs from such fonts more easily by using just ASCII
characters.