Delete pages not just from the page tree but also from the document
If a page is deleted, it is deleted from the page tree with /Parent set
to nil and returned.

This is problematic when deleting pages from a document and then saving
it because the page itself is still an indirect object of the document,
albeit invalid.

There is probably no reason to return the page object and to not delete
it from the document. Therefore it will now be deleted from the document
as well.