Extend and rename widget annotation button style to include color/size
Acrobat uses the /DA key in an undocumented way to allow users to define
the size and the color of the marker inside a check box or radio button.
The marker itself is defined by the /CA key in the appearance
characteristics dictionary.

To make use of this feature, the #button_style method is renamed to
\#button_marker_style and now only usable for check boxes and radio
buttons. It can return a value object holding all the information or be
used to set the marker, its size and its color.

The AppearanceGenerator makes use of this information when generating
the appearances.