Update AcroForm::ButtonField#create_widget for use with radio buttons
A radio button widget is one of a number of buttons of a radio button
field. Each of these widgets needs a (usually unique) value so that it
can be identified. This value is also used as field value to describe
the radio button widget that should be on.

This is done by using the value as name of an appearance state.