Fix parsing of long numbers when reading from IO
The tokenizer makes sure that 20 bytes are available when parsing a
token. This is okay for most cases.

However, if there is a long number and the characters of the number are
split due to not enough bytes having been prepared for the string
scanner, only the available characters are used for the number.

Fix this by making sure that all characters of the number are read
before converting the string into an integer or float.