Reconstruct cross reference table for damaged PDFs
Some PDF files are damaged in a way that preserves their content but
still makes them invalid. This happens, for example, when content is
added to the front or end of a file.

Most of these problems can be overcome easily be being more relaxed with
parsing, e.g. searching for the PDF marker in the first 1024 bytes.

Some of these errors are more serious because there is some invalid
information, e.g. wrong offsets in the cross reference table. Such an
error can still be recovered from by serially parsing the file front to
back and storing the found objects in a reconstructed cross refernce
table.

This commit implements such an algorithm and allows HexaPDF to read
nearly all PDF files.

There are two caveats for this:

1. The damaged file must be written in a way that allows such a
   recovery.

2. The reconstructed file may, in rare cases, not represent the original
   file correctly because some information is interpreted wrongly.