Update document/object validation code
The validation methods Object#validate and Document#validate do nearly
the same, yet yield different arguments to the provided block.
Furthermore, only the first validation problem can be reported.

This changes the internals so that regardless of which method is
invoked, the same kind of arguments are always yielded.

Furthermore, the internals of the Object#validate and
\#perform_validation methods has been changed so that multiple
validation problems may be yielded.

Since the interfaces and internal workings have been changed, this is
also a breaking change.