Fix infinite loop when reconstructing the cross-reference table
In special circumstances the cross-reference table reconstruction can
run into an infinite loop. This happens when a line starts with a string
that is not closed and exceeds the tokenizer's internal read buffer.

The bug is due to the next-new-line code not updating the tokenizer's
internal read buffer position because of an invalid regexp together with
tokenzier.scan_until. The regexp has \z in it and when the internal
buffer is empty (happens when the set position is not within the
internal read buffer), this just returns, making the next new line pos
the same as the current one.

Removing the \z fixes the problem. The additional consumation of
whitespace also updates the internal read buffer position and avoids
double scanning due to line changes.