Update cross-reference table parsing to handle invalidly numbered main sections
There are PDFs out in the wild which create the main cross-reference
section with wrong numbering while the rest of the PDF is fine.

The most often encountered problem is starting at 1 instead of 0:

xref
1 5
0000000000 65535 f
0000000011 00000 n
0000000022 00000 n
0000000033 00000 n
0000000044 00000 n

In this case the object with oid=2 at position 11 is actually the object
with oid=1 (and similar for the other objects).

By recognizing this problem and renumbering the OIDs in the
cross-reference section, we can correct the problem.