Fix Dictionary to allow adding new key-value pairs during validation
The 'hexapdf info -c' command fails for some sample PDFs because:

* The validation block checks for an object's type using #type
* Dictionary#type accesses #[](:type)
* Dictionary#[] populates the :type field with the default value,
  modifying the underlying hash value

This happens when there is a validation error for a set key, the /Type
field isn't set, and no class-level type is available.

To fix this we need to make sure that hash modifications are allowed
during iteration by using #keys.each instead of #each_key.