Change TrueType subsetter to avoid generating characters that need to be escaped
If text is output into a content stream, it needs to be serialized
according to the PDF spec rules. For serializing strings HexaPDF has
chosen the literal string syntax because the hexadecimal string syntax
would need twice as many characters.

For literal strings there are four characters -- \r, (, ), \ -- that
need to be escaped. If any of theses characters appears in the string,
String#gsub! has to actually do some work modifying the string.

By making sure that the encoded text using a TrueType font doesn't
contain those characters we avoid this work.

This change saves about 31% of object allocations and about 5% runtime
on the raw_text 10x TrueType benchmark.