Avoid parsing TTF post subtable if not necessary
The TrueType Post table contains information needed when using a
TrueType font. However, it also contains a subtable with glyph names
that is *not* used by HexaPDF but is automatically parsed.

By avoiding to parse this table until its information is needed we
reduce object allocations as well as enhance performance for the usual
case.

This is especially visible in long-lived processes that create a lot
of (small) documents.

The effect of the change was tested with the following script:

~~~
require 'hexapdf'

font_path = ARGV[0]

1000.times do
  doc = HexaPDF::Document.new
  font = doc.fonts.add(font_path)
  doc.pages.add.canvas.
    font(font, size: 10).
    text("The quick brown fox jumps over the lazy dog", at: [200, 200])
  doc.write('/tmp/out.pdf')
end
~~~

Results:

|        | Total allocated objects | Runtime |
|--------+-------------------------+---------|
| before |               2,672,939 |  4,267s |
| after  |               2,232,955 |  3,422s |
|--------+-------------------------+---------|
| change |        -14% / - 439,984 |    -20% |