Change Importer internals to better allow reclaiming documents
There are problems with memory retention when importing objects from
another document using the Importer. It seems that the documents are
never reclaimed by the garbage collector and that therefore the internal
hash used for caching grows.

I'm not quite sure why this is happening. However, looking at an
Importer instance we see that @source and @destination are wrapped by a
NullableWeakRef. However, the values of the @mapper hash are not wrapped
and store objects directly. Any all such objects contain a reference to
the document.

To avoid this dependency the mapped objects are now also encapsulated by
NullableWeakRef instances.