Fix infinite recursion when serializing a stream as part of a self-referencing object
The serializer uses an internal flag to avoid recursion. This flag is
used when serializing an indirect object and when serializing a stream.

An infinite recursion still occurs when an indirect object refers to a
stream and to itself (in this order!). The reason for this is that the
flag is reset when serializing a stream despite the stream not being the
main object being serialized. Later the self-reference of the object is
serialized and since the flag is not set, it is assumed that this is the
first invocation which means complete serialization again instead of
just outputting a reference.

The fix is to only reset the internal flag in the stream serialization
method if the stream itself is the main object.