Fix xref reconstruction to avoid O(n^2) performance problem
If a damaged file contains continuous lines starting with a '(' the
current algorithm will scan to the end of the file for each such line
before realizing that this is not a literal string but binary data.

This is a problem if there are *many* such lines.

Fix this problem by introducing a custom tokenization method that may
only be used during xref reconstruction because it just knows about the
two types of tokens (integers, keywords) that are relevant for it.