Fix xref reconstruction to try main xref table as last resort
If no trailer keyword is found, an error is thrown as the reconstructed
revision would be worthless without that information.

However, there might be cases where the trailer keyword doesn't start on
a new line. And even if that were not a problem, newer PDF files use
cross-reference streams which don't use the trailer keyword.

To better handle these cases, try, as a last resort, the cross-reference
table at the startxref location.