Change how the appearance streams are accessed
The current way of accessing an appearance stream by traversing the
needed objects is not very convenient and needs to be repeated
everywhere.

By implementing the #appearance method on Annotation to directly return
the requested appearance stream, this is simplified. Since #appearance
was used for something else before, this is a breaking change. This
functionality is now available through #appearance_dict.