Add Revision#update for updating the stored wrapper class
When an object is loaded from a PDF file, the wrapper class is
determined solely by the value of the object. This normally works
correctly but sometimes fails due to missing or invalid data.

This method allows updating the stored wrapper object so that the
correct wrapper object is later retrieved (e.g. during validation).