Delete invalid FontWeight values in Type::FontDescriptor
Some font descriptors in PDF files don't contain a valid font weight
value, e.g 350. This leads to a hard validation error.

However, since this value is optional and can be retrieved from the font
file, it is better to delete the invalid value instead of throwing a
hard validation error.