Fix xref reconstruction wrt to how linearized files are detected
There are invalid PDF files that actually contain two slightly different
versions of the same file. In such a case the current code invalidly
detects a linearied file which results in the trailer of the first
document to be used but the objects of the second.

Fix detection of linearized files to handle such cases.