Enhance AcroForm fallback font handling by allowing a callable object
Currently it is only possible to define a single fallback font for all
fields of an interactive form. This is a problem if the form uses
multiple different fonts that aren't usable by HexaPDF.

By allowing a callable object as the value of the
acro_form.fallback_font option it is possible to handle fallback fonts
based on the field and/or font object.

One possibly use case would be looking up the font name in the list of
operating system fonts.