Fix handling of non document-level, unencrypted metadata streams
According to the spec, the /EncryptMetadata entry should only apply to
the document-level metadata stream. However, in practice PDF creation
libraries and applications (Acrobat Distiller being one of them) won't
encrypt any metadata stream if /EncryptMetadata is set to false.

So we won't touch any metadata stream if /EncryptMetadata is set to
false.