Making sorting of segement infos for WidthFromPolygon more stable
Sorting is currently done by only using the minimum x value. So if two
entries have the same minimum x value the sort order may be Ruby
implementation dependent. By also using the maximum x value this can be
prevented.

We can also just sort on the whole sub array [min, max, stat] because
the first two items [min, max] won't be the same for any two sub arrays.

Example polygon where this is a problem, with corners (0, 0), (0, 10),
(5, 9), (10, 10), (10, 0):

    |\  /|
    | \/ |
    |    |
    |____|

If we want the width specification for the top with a height of 2 we
would get segement infos like

    [0, 0, :crossed_both]
    [-0.0, 10.0, :crossed_top]
    [10, 10, :crossed_both]

Before the first two elements might have been swapped depending on the
Ruby implementation. This is now prevented.