Update handling of AcroForm check boxes to allow widgets with different values
Check boxes can have multiple widgets and each widget can have a
different key for the "on state" of the appearance stream. This means
that selecting a check box is sometimes not possible in a simple
true/false manner.

To accomodate this check boxes are now handled nearly the same as radio
buttons. This means each widget may have a different key for the "on
state" appearance. And setting a check box can now be done via this key
or, in case of only one widget, with the true value.

This new behaviour is actually something that is actively used in the
wild because the no_toggle_to_off flag for radio buttons is not
implemented in any viewer. To get this behaviour one has to use a single
check box with multiple widgets with different values. This essentially
makes a radio button group where no button may be selected.